refactor: 탑승객 이름 추가
