+++ reservation-service/src/test/http/reservation-service.http
@@ -11,6 +11,7 @@ X-User-ID: 9b7f421f-47b0-47f5-aa5c-8f489697a57d
       "seatClass": "ECONOMY",
       "price": 120000,
       "passenger": {
+        "name": "홍길동",
         "birth": "1995-03-15",
         "gender": "MALE",
         "passportNumber": "M12345678"
@@ -21,6 +22,7 @@ X-User-ID: 9b7f421f-47b0-47f5-aa5c-8f489697a57d
       "seatClass": "BUSINESS",
       "price": 320000,
       "passenger": {
+        "name": "홍길동2",
         "birth": "1993-08-21",
         "gender": "FEMALE",
         "passportNumber": "F98765432"
@@ -65,6 +67,7 @@ Content-Type: application/json
     {
       "ticketId": "d5f61238-aef8-4d35-8d5f-bc4b130e4297",
       "passenger": {
+        "name": "홍길동",
         "birth": "1995-06-10",
         "gender": "FEMALE",
         "passportNumber": "P87654321"
